refactor(h2r): move filterAttributes to parseAndAdapt
